pyproject.toml
[tool.poetry]
name = "graph-recommendations-books"
version = "0.1.0"
description = ""
authors = ["Bartosz Mielczarek <bartoszmielczarekk@gmail.com>"]
readme = "README.md"
packages = [{include = "graph_recommendations_books"}]

[tool.poetry.dependencies]
python = "~3.10"
pyg-lib = {url = "https://data.pyg.org/whl/torch-1.13.0%2Bcpu/pyg_lib-0.1.0%2Bpt113cpu-cp310-cp310-linux_x86_64.whl"}
torch-cluster = {url = "https://data.pyg.org/whl/torch-1.13.0%2Bcpu/torch_cluster-1.6.0%2Bpt113cpu-cp310-cp310-linux_x86_64.whl"}
torch-scatter = {url = "https://data.pyg.org/whl/torch-1.13.0%2Bcpu/torch_scatter-2.1.0%2Bpt113cpu-cp310-cp310-linux_x86_64.whl"}
torch-sparse = {url = "https://data.pyg.org/whl/torch-1.13.0%2Bcpu/torch_sparse-0.6.15%2Bpt113cpu-cp310-cp310-linux_x86_64.whl"}
torch-spline-conv = {url = "https://data.pyg.org/whl/torch-1.13.0%2Bcpu/torch_spline_conv-1.2.1%2Bpt113cpu-cp310-cp310-linux_x86_64.whl"}
numpy = "~1.23.5"
pandas = "~1.5.2"
jupyter = "~1.0.0"
jupyterlab = "~3.5.0"
neo4j = "~5.3.0"
sentence-transformers = "~2.2.2"
torch = "1.13.0"
torch-geometric = "~2.2.0"
mlflow = "~2.0.1"
scikit-learn = "~1.2.0"
python-dotenv = "~0.21.0"
loguru = "^0.6.0"
pytorch-lightning = "1.8.1"
protobuf = "~3.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
